requirement.txt
ultralytics==8.0.199
Pillow==9.5.0